Fix styling for grid TOC
