--- theme/src/components/GridTableOfContents/GridTableOfContents.module.scss
@@ -1,6 +1,0 @@
-@import '@/scss/utils';
-
-.itemDimensions {
-  min-width: calculate-rem(73px, $min-font-size);
-  max-width: calculate-rem(115px);
-}
