--- theme/src/components/GridTableOfContents/GridTableOfContents.module.scss.d.ts
@@ -1,9 +1,0 @@
-export type Styles = {
-  itemDimensions: string;
-};
-
-export type ClassNames = keyof Styles;
-
-declare const styles: Styles;
-
-export default styles;
