--- theme/src/components/GridTableOfContents/index.ts
@@ -1,1 +1,0 @@
-export * from './GridTableOfContents';
